@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,13 +27,21 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  cupertino_icons: ^1.0.2
+  flutter_localizations:
+    sdk: flutter
   flutter:
     sdk: flutter
+  cupertino_icons: ^1.0.2
+  firebase_auth: ^4.2.6
+  firebase_core: ^2.5.0
   flutter_bloc: ^8.1.2
-  flutter_blue_plus: ^1.4.0
-  flutter_bluetooth_serial: ^0.4.0
+  flutter_blue: ^0.8.0
+  flutter_screenutil: ^5.6.1
+  flutter_svg: ^2.0.1
   system_settings: ^2.1.0
+  intl: ^0.17.0
+  fluttertoast: ^8.1.3
+  google_sign_in: ^5.4.4
 dev_dependencies:
 
   # The "flutter_lints" package below contains a set of recommended lints to
@@ -54,6 +62,34 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+  assets:
+    - assets/images/
+    - assets/images/svg/
+    - lang/
+  fonts:
+    - family: Inter
+      fonts:
+        - asset: assets/fonts/Inter/Inter-Thin.ttf
+          weight: 100
+        - asset: assets/fonts/Inter/Inter-ExtraLight.ttf
+          weight: 200
+        - asset: assets/fonts/Inter/Inter-Light.ttf
+          weight: 300
+        - asset: assets/fonts/Inter/Inter-Regular.ttf
+          weight: 400
+        - asset: assets/fonts/Inter/Inter-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Inter/Inter-SemiBold.ttf
+          weight: 600
+        - asset: assets/fonts/Inter/Inter-Bold.ttf
+          weight: 700
+        - asset: assets/fonts/Inter/Inter-ExtraBold.ttf
+          weight: 800
+        - asset: assets/fonts/Inter/Inter-Black.ttf
+          weight: 900
+    - family: DG-3asomy
+      fonts:
+        - asset: assets/fonts/DG-3asomy-Regular.ttf
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
