@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:sense_more/business_logic/helpers/bluetooth_helper.dart';
+import 'package:sense_more/core/shared/bluetooth_helper.dart';
 
 part 'log_state.dart';
 
